+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Deprecated
 ### Removed
+
+- The long deprecated union types have been removed, in favor of the new sum
+  types (#1245).
+
 ### Fixed
 ### Security